pyproject.toml
[tool.poetry]
name = "youtube-analytics-project1"
version = "0.1.0"
description = ""
authors = ["Anastasiya Brodskaya <anastasia.brodskaja@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
